Templatize bfirst_search_init() and bfirst_search_try() so they can be used with types other than int16_t
git-svn-id: https://svn.eduke32.com/eduke32@7629 1a8010ca-5511-0410-912e-c29ae57300e0

# Conflicts:
#	source/build/src/build.cpp

+++ source/duke3d/src/astub.cpp
@@ -6718,7 +6718,7 @@ static void Keys3d(void)
 
             static int16_t sectlist[MAXSECTORS];
             static uint8_t sectbitmap[(MAXSECTORS+7)>>3];
-            int32_t sectcnt, sectnum;
+            int16_t sectcnt, sectnum;
 
             i = searchsector;
             if (CEILINGFLOOR(i, stat)&1)
